// autogpt_platform/frontend/src/app/api/__generated__/models/myUnpublishedAgent.ts
/**
 * Generated by orval v7.13.0 🍺
 * Do not edit manually.
 * AutoGPT Agent Server
 * This server is used to execute agents that are created by the AutoGPT system.
 * OpenAPI spec version: 0.1
 */
import type { MyUnpublishedAgentAgentImage } from "./myUnpublishedAgentAgentImage";
import type { MyUnpublishedAgentRecommendedScheduleCron } from "./myUnpublishedAgentRecommendedScheduleCron";

export interface MyUnpublishedAgent {
  graph_id: string;
  graph_version: number;
  agent_name: string;
  agent_image?: MyUnpublishedAgentAgentImage;
  description: string;
  last_edited: Date;
  recommended_schedule_cron?: MyUnpublishedAgentRecommendedScheduleCron;
}