// autogpt_platform/frontend/src/app/api/__generated__/models/myUnpublishedAgentsResponse.ts
/**
 * Generated by orval v7.13.0 🍺
 * Do not edit manually.
 * AutoGPT Agent Server
 * This server is used to execute agents that are created by the AutoGPT system.
 * OpenAPI spec version: 0.1
 */
import type { MyUnpublishedAgent } from "./myUnpublishedAgent";
import type { Pagination } from "./pagination";

export interface MyUnpublishedAgentsResponse {
  agents: MyUnpublishedAgent[];
  pagination: Pagination;
}