// autogpt_platform/frontend/src/app/api/__generated__/models/storeSubmissionsResponse.ts
/**
 * Generated by orval v7.13.0 🍺
 * Do not edit manually.
 * AutoGPT Agent Server
 * This server is used to execute agents that are created by the AutoGPT system.
 * OpenAPI spec version: 0.1
 */
import type { StoreSubmission } from "./storeSubmission";
import type { Pagination } from "./pagination";

export interface StoreSubmissionsResponse {
  submissions: StoreSubmission[];
  pagination: Pagination;
}